Refactor profile stuff into own file
